requirements.txt
fastapi==0.88.0
uvicorn==0.22.0
pyyaml
python-multipart
tika
pdfminer
numpy
pdf2image
opencv-python
pdfplumber
mysql-connector-python==9.2.0
bcrypt==4.2.1
# brew install poppler
google-generativeai